backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic>=2.6.0
pydantic-settings>=2.2.0
anthropic>=0.40.0
google-genai>=1.7.0
python-dotenv>=1.0.0
httpx>=0.27.0
pandas>=2.0.0
plotly>=5.0.0
mcp>=1.0.0
snowflake-connector-python>=3.0.0
